Relicense reference implementation to 3-clause BSD.
The prose should be enough to show someone 'skilled in the art' how
to implement the idea.  However, a common criticism of patents is
the lack of an example reference implementation to make the idea
clear to a programmer.

License the example code in a way which permits re-use as a starting
point for another implementation in many software-license-environments.